Replace constructor code with first-turn init
It appears that when the MyRobot constructor is called, none of
the inherited variables have yet been initialised, meaning that
they are all null pointers. Any attempt to initialise MyRobot
using those inherited variables will therefore crash.

To fix this initialisation has been moved to the start of the
turn function, where it is guaranteed that all of the required
variables already exit.