Optimise castle-talk data model
In the past data was interpreted on a bit-by-bit basis, meaning that
castle-talk could transmit 8 pieces of data.

Instead, it is now interpreted on a number-by-number basis. With 8
bits, castle-talk can transmit 256 different types of data. Reserving
the zero for a no-op-style message, this leaves 255 different types
of messages that can be sent. This leaves no bits wasted.