Fix unsynced reserve values, centralise pilgrims
Resource reserve values were previously out-of-sync due to the
turncount being inaccurate for churches. This is fixed by ensuring that
all structures and pilgrims are always aware of the global turn (called
builderTurn), and that this information is propagated via broadcasts.
Signal radii are also used in this as up to 22 bits of data may need to
be transmitted. The datapack is in the following format:

most sig 1XXX XXXX XXXX XXXX least sig
          ABC |__location__|
where ABC denotes the three least-significant bits of the turn. The
remainder of the turn (i.e. the 7 most-significant bits) are
transmitted in the signal radius, but with an increment of 3 in order
to differentiate it from regular assignment signals.

As a consequence pilgrim construction must also be centralised in order
to ensure that they can simply read the global turn off the castle that
built them. Therefore invasions to the enemy side are now managed by
castles, entirely.